+++ emirdrp/ext/gtc/__init__.py
@@ -4,10 +4,10 @@ try:
 
     from numina.ext.gtc import register
 
-    _more = {'emirdrp.dataproducts.MasterBias': DF.TYPE_FRAME,
-             'emirdrp.dataproducts.MasterDark': DF.TYPE_FRAME,
-             'emirdrp.dataproducts.MasterIntensityFlat': DF.TYPE_FRAME,
-             'emirdrp.dataproducts.CoordinateList2DType': DF.TYPE_DOUBLE_ARRAY2D,
+    _more = {'emirdrp.products.MasterBias': DF.TYPE_FRAME,
+             'emirdrp.products.MasterDark': DF.TYPE_FRAME,
+             'emirdrp.products.MasterIntensityFlat': DF.TYPE_FRAME,
+             'emirdrp.products.CoordinateList2DType': DF.TYPE_DOUBLE_ARRAY2D,
              }
     register(_more)
 except ImportError:
