+++ emirdrp/recipes/aiv/mask.py
@@ -30,9 +30,9 @@ from numina.core import Requirement, Product, Parameter
 from numina.core.requirements import ObservationResultRequirement
 from numina.constants import FWHM_G
 from emirdrp.core import EmirRecipe
-from emirdrp.dataproducts import DataFrameType
-from emirdrp.dataproducts import CoordinateList2DType
-from emirdrp.dataproducts import ArrayType
+from emirdrp.products import DataFrameType
+from emirdrp.products import CoordinateList2DType
+from emirdrp.products import ArrayType
 from emirdrp.requirements import MasterBiasRequirement
 from emirdrp.requirements import MasterDarkRequirement
 from emirdrp.requirements import MasterIntensityFlatFieldRequirement
