+++ emirdrp/recipes/engineering/detectorgain.py
@@ -34,7 +34,7 @@ from numina.core import Product
 from emirdrp.core import EmirRecipe
 from emirdrp.instrument.detector import QUADRANTS
 from emirdrp.instrument.channels import CHANNELS
-from emirdrp.dataproducts import MasterGainMap, MasterRONMap
+from emirdrp.products import MasterGainMap, MasterRONMap
 
 _logger = logging.getLogger('numina.recipes.emir')
