+++ emirdrp/recipes/image/dither.py
@@ -37,7 +37,7 @@ from emirdrp.requirements import Offsets_Requirement
 from emirdrp.requirements import Catalog_Requirement
 from emirdrp.requirements import SkyImageSepTime_Requirement
 
-from emirdrp.dataproducts import SourcesCatalog
+from emirdrp.products import SourcesCatalog
 
 from .shared import DirectImageCommon
