+++ emirdrp/recipes/image/shared.py
@@ -55,7 +55,7 @@ from numina.util.sextractor import SExtractor
 from numina.util.sextractor import open as sopen
 import numina.util.sexcatalog as sexcatalog
 
-from emirdrp.dataproducts import SourcesCatalog
+from emirdrp.products import SourcesCatalog
 
 from .checks import check_photometry
 from .naming import (name_redimensioned_frames, name_object_mask,
